Refactor authentication package to auth and adjust related imports and usages.
